requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
PyPDF2>=3.0.0
flask>=2.3.0
flask-cors>=4.0.0
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0 